feat(mcp): Stage 138 — add memory and template preview wrappers
Move agent-tool-recommendation-memory + template-effectiveness-signals
into @conclave-ai/workspace-preview with subpath exports; dashboard files
become thin re-export wrappers (zero call-site churn). Moved their tests
too. All Stage 126~129 derived moat-signal helpers now live in the shared
package (one source of truth).

MCP Basic snapshot wrappers: previewAgentToolMemory + previewTemplateSignals
— pure, read-only, return {ok,kind,preview} + boundary (mutatesState/
usesHostedExecution/requiresPayment false, derivedPreviewOnly true);
malformed input does not throw. The preview wrapper module now covers all
8 preview tools. No server runtime wiring, no network/env/mutation/LLM/
payment/agent/PR-post.

workspace-preview 175/175, mcp-workspace 41/41, dashboard 218/218
(2 helper suites moved to package), monorepo typecheck 57/57, build green.
